pyproject.toml
[tool.poetry]
name = "jaspelku-fastapi"
version = "0.1.0"
description = ""
authors = ["arlianasrori <arlianasrori@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.1"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.31"
alembic = "^1.13.2"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
pydantic = "^2.8.2"
python-random-strings = "^0.3"
asyncpg = "^0.29.0"
psycopg2 = "^2.9.9"
bcrypt = "^4.2.0"
google-auth = "^2.33.0"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
